Fix linker error from libprojectM pkg-config bug
libprojectM 4.1.6 pkg-config outputs "-l:projectM-4" but the library
uses standard naming (libprojectM-4.so), requiring "-lprojectM-4".
The "-l:" syntax makes the linker search for an exact filename.

Add sed filter to strip the errant colon. This is a no-op if upstream
fixes the bug, so no future Makefile changes will be needed.